[reporting] Move reporting under sos/report
Moves the HTML/JSON reporting bits under sos/report.
